refactor(drag-to-create): address review (drop unused export, split rect util, simplify onSelect)
- Remove the public DragSelection export; it's an internal computeRange
  return type, not used by any public signature.
- Move Rect/unionRect/intersectRect (+ specs) out of selection.ts into
  utils/rect.ts, matching the one-concern-per-util-file layout of sibling
  plugins; selection.ts is now just selection semantics.
- Drop the one-field DragToCreateContext wrapper; onSelect now takes
  (selection, openEventForm) directly.
