+++ docs/logs/2026-06-21.md
@@ -15,6 +15,7 @@
 - **[drag-to-create / fix]**: Two mirror-overlay bugs found in the demo (short "Extra Small 300px" calendar). (1) The selection overlay painted far outside the calendar body: the mirror was the union of the start/last cell `getBoundingClientRect()`s, which extend past the clipped (scrolled) calendar body, and the overlay is portaled to `document.body` so the body's overflow can't clip it. Fix: clip the mirror to the `[data-testid="calendar-body"]` ancestor rect. Extracted pure `unionRect(Rect, Rect)` (moved off the old element-based version) and new `intersectRect(Rect, Rect): Rect | null` into `selection.ts`; the provider's `computeMirror` reads the DOMRects and clips. 7 new pure specs in `selection.test.ts` (incl. the exact overflow case, full-edge clip, and zero-area edge-touch → null). (2) Native text selection highlighted event titles/day-numbers during a drag. Fix: on gesture activation, set `user-select: none` (+ `-webkit-user-select`) on `document.body` and clear any in-progress selection; restored in `reset()` (covers commit/cancel/blur/Escape). 31 tests in the package. Note: clipping couples to the existing `data-testid="calendar-body"` marker (the visible calendar bounds for both regular and resource calendars).
 - **[drag-to-create / core fix]**: Feature didn't work in vertical daily-granularity views (resource month vertical, week at daily granularity). Root cause: `vertical-grid-col.tsx` passed `hour={day.hour()}` for ALL cells, so even a `gridType="day"` cell described a 00:00-01:00 one-hour range instead of a full day. That made `isFullDaySpan` false, so the plugin's cross-day clamp rejected every other cell and the selection never extended. The horizontal grid already does `hour={gridType === 'hour' ? day.hour() : undefined}` (which is why month worked). Fix: mirror that conditional in `vertical-grid-col.tsx` so daily cells emit a full-day range. Side benefit: click-create on a daily vertical cell now makes a full-day event (consistent with horizontal month) instead of a midnight 1-hour event. TDD: new `vertical-grid-col.test.tsx` case asserts a `gridType="day"` cell's `data-end - data-start` is 23h (RED before, GREEN after); also set the test helper's default `gridType: 'hour'` so the existing hourly/sub-hour tests still exercise hourly behavior. Full calendar suite 795 pass.
 - **[drag-to-create / review fixes]**: Code-review cleanup. Removed two leftover debug `console.log`s from the demo-debugging session (`use-calendar-navigation.ts`, `droppable-cell.tsx` — both ran on every render). Replaced the mirror-clip's `[data-testid="calendar-body"]` selector with a stable semantic marker: added `data-calendar-viewport="true"` to the calendar body in `ilamy-calendar.tsx` (test id kept) and pointed `computeMirror` at `[data-calendar-viewport]`, so a test-id rename can't silently break the overlay clipping. Fixed a stale `dragcreate` → `drag-to-create` comment in the plugin's `bunup.config.ts`. Calendar 795 + plugin 31 tests pass; type-check + biome clean (one pre-existing, unrelated `noConsole`-suppression warning left as-is).
+- **[drag-to-create / refactor review]**: Post-review cleanup of PR #214. (1) Removed the unused public `DragSelection` export from the plugin index, it's the internal return of `computeRange` and no public signature used it (siblings only export consumer-facing types). (2) Split the mirror rect geometry (`Rect`/`unionRect`/`intersectRect`) out of `selection.ts` into `utils/rect.ts` + `rect.test.ts`, so `selection.ts` is just selection semantics and the layout matches the agenda plugin's one-concern-per-util-file convention. (3) Dropped the one-field `DragToCreateContext` wrapper; `onSelect` now takes `(selection, openEventForm)` directly. 31 plugin tests pass (now across 4 files); calendar + plugin type-check and biome clean.
 - **[docs / findings]**: Documented the configured-dayjs zone-carrying-string bug in `docs/configured-dayjs-timezone-bug.md` — root cause (the `dayjs()`→`dayjs.tz()` alias + how `dayjs.tz` re-stamps a string's literal digits as the target tz, per the plugin source), the safe `dayjs.utc(iso).tz(tz)` pattern, and the verified library-wide affected-site list (consumer event ingestion in `normalize.ts:40-41`; recurrence `update-recurring-event.ts:69/256` and `ical.ts` EXDATE/RECURRENCE-ID; `generate-recurring-events.ts` until). Confirmed it's an **upstream dayjs bug** (iamkun/dayjs#2946, open; proposed PR #2949) that our wrapper inherits library-wide; the upstream workaround `dayjs(str).tz(tz)` doesn't work for us because our `dayjs()` is already `dayjs.tz()`. **Decision: the library-wide fix is a separate effort, not part of this PR** — drag-to-create works around it locally via `read-cell`'s `dayjs.utc().tz()`. The demo/tests build event dates as Dayjs/Date so the consumer-ingestion path isn't exercised today, but the public `start: Dayjs | Date | string` contract leaves it exposed.
 
 ## Files Modified
