+++ packages/plugins/drag-to-create/src/index.ts
@@ -1,4 +1,3 @@
 // Public entry for `@ilamy/calendar-drag-to-create`.
 export { dragToCreatePlugin } from './drag-to-create-plugin'
-export type { DragToCreateContext, DragToCreateOptions } from './types'
-export type { DragSelection } from './utils/selection'
+export type { DragToCreateOptions } from './types'
